added logout, token refresh, privalege control

+++ resources/item.py
@@ -1,5 +1,11 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required, 
+    get_jwt_claims, 
+    jwt_optional, 
+    get_jwt_identity, 
+    fresh_jwt_required
+)
 from models.item import ItemModel
 
 class Item (Resource):
@@ -10,13 +16,14 @@ class Item (Resource):
                         help="This field cannot be left blank")
 
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {"message":"item not found"},404
 
+    @fresh_jwt_required
     def post(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -31,7 +38,7 @@ class Item (Resource):
 
 
 
-
+    @jwt_required
     def delete (self, name):
         # connection = sqlite3.connect("data.db")
         # cursor = connection.cursor()
@@ -40,6 +47,10 @@ class Item (Resource):
         # connection.commit()
         # connection.close()
         # return {"message":"item'{}' deleted".format(name)}
+
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"message":"admin previlege required"}
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -61,6 +72,7 @@ class Item (Resource):
 
 
 class ItemList (Resource):
+    @jwt_optional
     def get (self):
         # item_list= ItemModel.query.all()
         # res = []
@@ -70,5 +82,11 @@ class ItemList (Resource):
         # or
         # res = list(map(lambda x: x.json(), ItemModel.query.all()))
         # or
-        res = [item.json() for item in ItemModel.find_all()]
-        return {'items':res}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items':items}
+        return {
+            "items": [item["name"] for item in items],
+            "message":"log in for item details"
+        },200
